Fix 20 compilation warnings across decoder, rendering, audio, and emulation
- [-Wimplicit-fallthrough]: Mark 3 intentional switch fallthroughs in decoder.cpp
  (PRM_FRN_SZ→FRN, PRM_FRM_SZ→FRM, NDO_Delayslot→NDO_NextOp) with FALLTHROUGH macro.
  Added portable FALLTHROUGH macro to types.h (GCC __attribute__, C++17 [[fallthrough]], fallback).

- [-Wcast-function-type]: New TA_PP_FP() macro in ta.h routes function pointer casts
  via void(*)(void) (GCC's universal-compat type) instead of direct type reinterpretation.
  Applied to 7 AppendPolyParamN entries in ta_const_df.h and ta.h FifoSplitterEx copy.

- [-Wempty-body]: Changed disabled printf in cdi.cpp from empty to ((void)0) expression;
  fixes 4 if-statements that had no body. Added comment explaining the idiom.

- [-Wunused-but-set-parameter]: SetupPaletteForTexture palette_index now has (void) cast
  and a comment noting CLUT upload is not yet wired to GX.

- [-Wtype-limits]: iso.cpp _ISO9660_read_r had a dead signed-check on unsigned 'len'.
  Real bug fix: hold _read() result in signed int rlen, test it, then assign to len.

- [-Wextra enumerated/non-enumerated ternary]: 5 sites in wii_driver.cpp now cast
  the enum branch to u32 so both ternary arms are the same type.

No behavioral changes except the iso.cpp error-handling fix.

+++ wii/dc/sh4/rec_v2/wii_driver.cpp
@@ -680,7 +680,7 @@ void bop_resolve_a_d(shil_opcode* op)
 	}
 
 	ppc_ireg d=GetIntReg(op->rd._reg);
-	bop_d = (d!=ppc_rinvalid) ? (u32)d : ppc_rarg0;
+	bop_d = (d!=ppc_rinvalid) ? (u32)d : (u32)ppc_rarg0;
 }
 
 // Resolve rs2 -> bop_b (source). Materialises an immediate into rarg1.
@@ -786,7 +786,7 @@ void unop3_start(shil_opcode* op)
 	}
 
 	ppc_ireg d=GetIntReg(op->rd._reg);
-	bop_d = (d!=ppc_rinvalid) ? (u32)d : ppc_rarg0;
+	bop_d = (d!=ppc_rinvalid) ? (u32)d : (u32)ppc_rarg0;
 }
 
 void unop3_end(shil_opcode* op)
@@ -847,14 +847,14 @@ static void fbop_resolve(shil_opcode* op)
 	fop_a=fsrc_or_load(op->rs1,ppc_farg0);
 	fop_b=fsrc_or_load(op->rs2,ppc_farg1);
 	ppc_freg d=get_fpu_pin_preset()?GetFloatReg(op->rd._reg):ppc_finvalid;
-	fop_d=(d!=ppc_finvalid)?(u32)d:ppc_farg0;
+	fop_d=(d!=ppc_finvalid)?(u32)d:(u32)ppc_farg0;
 }
 static void fuop_resolve(shil_opcode* op)
 {
 	verify(!op->rs1.is_null() && !op->rd.is_null());
 	fop_a=fsrc_or_load(op->rs1,ppc_farg0);
 	ppc_freg d=get_fpu_pin_preset()?GetFloatReg(op->rd._reg):ppc_finvalid;
-	fop_d=(d!=ppc_finvalid)?(u32)d:ppc_farg0;
+	fop_d=(d!=ppc_finvalid)?(u32)d:(u32)ppc_farg0;
 }
 static void fbop_end(shil_opcode* op)
 {
@@ -3444,7 +3444,7 @@ DynarecCodeEntry* ngen_Compile(DecodedBlock* block,bool force_checks)
 				// Resolve dest: a pinned (integer) rd is written in place; a
 				// non-pinned or float-typed rd uses rarg0 + a context store.
 				ppc_ireg rdr = op->rd.is_r32i() ? GetIntReg(op->rd._reg) : ppc_rinvalid;
-				u32 dst = (rdr!=ppc_rinvalid) ? (u32)rdr : ppc_rarg0;
+				u32 dst = (rdr!=ppc_rinvalid) ? (u32)rdr : (u32)ppc_rarg0;
 
 				if (op->rs1.is_imm())
 				{
